+++ test/dart_autogui_test.dart
@@ -2,15 +2,17 @@ import 'package:dart_autogui/dart_autogui.dart';
 import 'package:test/test.dart';
 
 void main() {
-  group('A group of tests', () {
-    final awesome = Awesome();
-
-    setUp(() {
-      // Additional setup goes here.
+  group('Easing tests', () {
+    test('Linear easing', () {
+      expect(easeLinear(0.0), 0.0);
+      expect(easeLinear(0.5), 0.5);
+      expect(easeLinear(1.0), 1.0);
     });
 
-    test('First Test', () {
-      expect(awesome.isAwesome, isTrue);
+    test('Quad easing', () {
+      expect(easeInQuad(0.0), 0.0);
+      expect(easeInQuad(0.5), 0.25);
+      expect(easeInQuad(1.0), 1.0);
     });
   });
 }